Clean up Gaming.html
Gaming.html is being cleaned up for readability.

Not looking at the file for a few months and all I see is spaghetti code.

It's becoming much cleaner, proper indentation, redundant code is removed/moved to a new CSS file: style2.css.

In general, a few hours of cleaning the code up will go a long way.

There's just way too much stuff in the site folder that doesn't need to be in there.

Folder structures are confusing.  It all needs to be re-organized now that I know what content we need.