Removing Everything & Starting Over
Now that we have a clear idea how the site needs to be laid out, I'm removing most of the content (or moving it somewhere) so the site can be changed easier.

I'm moving all content into '/hockey' or '/game' folders, and treating each folder as a seperate "site".  So all images, videos, everything is in that related folder.

Second, I'm going from the top of the main page to the bottom and checking each line of code to see what is doing what, and what doesn't need to be there.

Is there a simpler way to do the same thing?

Yes. Most of the time it's yes.  So a good chunk of content is temporarily removed and I'm rebuilding the main page from the beginning.

Sections of the website are now using condensed names, like 'TV' instead of 'Television'.

Nav Pill Bars are now fixed to top, and act as the main point of navigation.

Removed the redundant "Portfolio", "About" and "Contact" links.

The about will be the first section of the website, and condensed (height-wise).

Contact will just be always visible on the bottom right of the page.

Main title is now called "Dave Whiffen's Portfolio" instead of just "Dave".

Nav Pills have been tweaked to use less code and follow unified classes.

Content has been temporarily removed to focus on the one "Rise and Fall" project and write the code cleaner and simpler.

All content is formatted properly with parents being shown.

Every line has been commented to explain what a line is doing.

The footer has been removed for now, since it wasn't functioning properly.

Room for a description of the game has been added.