Rise and Fall Commentary
Added photos, description and a step-by-step breakdown of the Level Design,
the game's core design philosophy and mechanics,
the Early Whitebox and Iterations
Unique systems we implemented based on playtesting and feedback

A closing statement

Added Headshots for every person on the team.
Cropped photos so the headshots are all relatively the same size.

Organized photos by skillset.

Built a headshots-sm class to keep the styling consistent.

Created a bright blue button at the end of the page.

Linked the button to the Github repository.

Cleaned up some CSS classes that were linking into other classes.

Created a new, cleaner, H1 header class.

Formatted important text in bold.

Framed screenshots to compliment what the text was talking about.

Created a custom CSS class for the photos and made sure they resize properly with the window.